add Free procedure for a Select_Statement fix bug from Rob Veenker [veenker@xs4all.nl] about bind parameters.
Text from Rob
---------------------------------------------
As I was using the ODBC binding (I realize it's a demo version) for calling
stored procedures in Oracle, I stubled on a problem. It seemed that the length
of a parameter value (using parameter binding) was incorrectly passed to
ODBC.

I noticed that te routine Databases.Execute used a newly declared Parameter
record for each parameter. However the last parameter in the call to
SQLBindParameter is used as a 'access variable and therefor the storage is
needed even after the binding has finished.

I solved the problem by using a temporary array (in stead of a single record)
to shorten the notation for the parameters of SQLBindParameter.
---------------------------------------------